added

+++ Microservice-SB-Ticket-Booking/user-service/src/main/java/com/learning/userdetails/repository/IUserRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IUserRepo extends JpaRepository<User,Long> {
+public interface IUserRepo extends JpaRepository<User,Integer> {
     User findByUserEmail(String email);
 }
